+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 *.db
 receipts/
 package-lock.json
+收據/
